feat(nx): add run-many
This adds a run-many command to:

 - make more semantic sense for the --all option
 - to add the ability to specify specific projects to run

It also involves a quite substancial refactoring of affected.ts in order
to attempt reuse of the code in both commands
