+++ packages/workspace/src/command-line/run-tasks/index.ts
@@ -1,0 +1,2 @@
+export * from './run-many';
+export * from './affected';
